Fix after update to nightly
