+++ org.metaborg.lang.tiger.interpreter/.gitignore
@@ -6,3 +6,10 @@ src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated
 src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/terms
 src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/terms/build
 src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/terms/match
+src/test/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/test
+/tiger.launch
+/Testtiger.launch
+/tiger (Daemon).launch
+/tiger-server
+/tiger-client
+/tiger
