+++ org.metaborg.lang.tiger.interpreter/pom.xml
@@ -32,6 +32,10 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
                   <includes>**/**</includes>
                 </fileset>
                 <fileset>
+                  <directory>src/test/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/test</directory>
+                  <includes>**/**</includes>
+                </fileset>
+                <fileset>
                   <directory>${basedir}</directory>
                   <includes>src/main/resources/specification.aterm</includes>
                 </fileset>
@@ -43,6 +47,14 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
             </configuration>
           </execution>
         </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>2.19.1</version>
+        <configuration>
+          <argLine></argLine>
+        </configuration>
       </plugin>
     </plugins>
   </build>
@@ -89,6 +101,11 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
       <groupId>org.metaborg</groupId>
       <artifactId>org.spoofax.terms</artifactId>
       <version>2.1.0-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>com.martiansoftware</groupId>
+      <artifactId>nailgun-server</artifactId>
+      <version>0.9.1</version>
     </dependency>
   </dependencies>
 </project>
