+++ org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/lang/tiger/interpreter/natives/eqI_2.java
@@ -1,9 +1,8 @@
-package tiger.interpreter.natives;
+package org.metaborg.lang.tiger.interpreter.natives;
 
-import org.metaborg.lang.tiger.interpreter.natives.concatS_2NodeGen;
+import org.metaborg.lang.tiger.interpreter.natives.eqI_2NodeGen;
 import org.metaborg.meta.lang.dynsem.interpreter.nodes.building.TermBuild;
 
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.NodeChildren;
 import com.oracle.truffle.api.dsl.Specialization;
@@ -11,25 +10,20 @@ import com.oracle.truffle.api.source.SourceSection;
 
 @NodeChildren({ @NodeChild(value = "left", type = TermBuild.class),
 		@NodeChild(value = "right", type = TermBuild.class) })
-public abstract class concatS_2 extends TermBuild {
+public abstract class eqI_2 extends TermBuild {
 
-	public concatS_2(SourceSection source) {
+	public eqI_2(SourceSection source) {
 		super(source);
 	}
 
 	@Specialization
-	public String doString(String left, String right) {
-		return doStringConcat(left, right);
-	}
-
-	@TruffleBoundary
-	private String doStringConcat(String left, String right) {
-		return left + right;
+	public int doInt(int left, int right) {
+		return left == right ? 1 : 0;
 	}
 
 	public static TermBuild create(SourceSection source, TermBuild left,
 			TermBuild right) {
-		return concatS_2NodeGen.create(source, left, right);
+		return eqI_2NodeGen.create(source, left, right);
 	}
 
 }
