+++ org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/lang/tiger/interpreter/natives/eqS_2.java
@@ -1,6 +1,6 @@
-package tiger.interpreter.natives;
+package org.metaborg.lang.tiger.interpreter.natives;
 
-import org.metaborg.lang.tiger.interpreter.natives.divI_2NodeGen;
+import org.metaborg.lang.tiger.interpreter.natives.eqS_2NodeGen;
 import org.metaborg.meta.lang.dynsem.interpreter.nodes.building.TermBuild;
 
 import com.oracle.truffle.api.dsl.NodeChild;
@@ -10,20 +10,20 @@ import com.oracle.truffle.api.source.SourceSection;
 
 @NodeChildren({ @NodeChild(value = "left", type = TermBuild.class),
 		@NodeChild(value = "right", type = TermBuild.class) })
-public abstract class divI_2 extends TermBuild {
+public abstract class eqS_2 extends TermBuild {
 
-	public divI_2(SourceSection source) {
+	public eqS_2(SourceSection source) {
 		super(source);
 	}
 
 	@Specialization
-	public int doInt(int left, int right) {
-		return left / right;
+	public int doInt(String left, String right) {
+		return left.equals(right) ? 1 : 0;
 	}
 
 	public static TermBuild create(SourceSection source, TermBuild left,
 			TermBuild right) {
-		return divI_2NodeGen.create(source, left, right);
+		return eqS_2NodeGen.create(source, left, right);
 	}
 
 }
