+++ org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/lang/tiger/interpreter/natives/notI_1.java
@@ -1,6 +1,6 @@
-package tiger.interpreter.natives;
+package org.metaborg.lang.tiger.interpreter.natives;
 
-import org.metaborg.lang.tiger.interpreter.natives.printS_1NodeGen;
+import org.metaborg.lang.tiger.interpreter.natives.notI_1NodeGen;
 import org.metaborg.meta.lang.dynsem.interpreter.nodes.building.TermBuild;
 
 import com.oracle.truffle.api.dsl.NodeChild;
@@ -8,26 +8,28 @@ import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.source.SourceSection;
 
 @NodeChild(value = "stringbuild", type = TermBuild.class)
-public abstract class exitI_1 extends TermBuild {
+public abstract class notI_1 extends TermBuild {
 
-	public exitI_1(SourceSection source) {
+	public notI_1(SourceSection source) {
 		super(source);
 	}
 
 	/**
-	 * Exit the program with exit code status.
+	 * Return 1 if boolean = 0, else return 0.
 	 * 
 	 * @param i
 	 * @return
 	 */
 	@Specialization
 	public int doInt(int i) {
-		System.exit(i);
-		
-		return 0;
+		if (i == 0) {
+			return 1;
+		} else {
+			return 0;
+		}
 	}
 
 	public static TermBuild create(SourceSection source, TermBuild stringbuild) {
-		return exitI_1NodeGen.create(source, stringbuild);
+		return notI_1NodeGen.create(source, stringbuild);
 	}
 }
