+++ org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/lang/tiger/interpreter/natives/printI_1.java
@@ -1,6 +1,6 @@
-package tiger.interpreter.natives;
+package org.metaborg.lang.tiger.interpreter.natives;
 
-import org.metaborg.lang.tiger.interpreter.natives.printS_1NodeGen;
+import org.metaborg.lang.tiger.interpreter.natives.printI_1NodeGen;
 import org.metaborg.meta.lang.dynsem.interpreter.nodes.building.TermBuild;
 
 import com.oracle.truffle.api.dsl.NodeChild;
@@ -8,19 +8,18 @@ import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.source.SourceSection;
 
 @NodeChild(value = "stringbuild", type = TermBuild.class)
-public abstract class printS_1 extends TermBuild {
+public abstract class printI_1 extends TermBuild {
 
-	public printS_1(SourceSection source) {
+	public printI_1(SourceSection source) {
 		super(source);
 	}
 
 	@Specialization
-	public String doString(String s) {
-		System.out.println(s);
-		return s;
+	public String doString(int i) {
+		return i + "";
 	}
 
 	public static TermBuild create(SourceSection source, TermBuild stringbuild) {
-		return printS_1NodeGen.create(source, stringbuild);
+		return printI_1NodeGen.create(source, stringbuild);
 	}
 }
