+++ org.metaborg.lang.tiger.interpreter/tiger.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.jdt.launching.localJavaApplication">
   <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-  <listEntry value="/org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/tigerLanguage.java"/>
+  <listEntry value="/org.metaborg.lang.tiger.interpreter/src/main/java/org/metaborg/org/metaborg/lang/tiger/interpreter/generated/tigerEntryPoint.java"/>
   </listAttribute>
   <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
   <listEntry value="1"/>
@@ -12,4 +12,5 @@
   <stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="${resource_loc}"/>
   <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="org.metaborg.lang.tiger.interpreter"/>
   <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.m2e.launchconfig.sourcepathProvider"/>
+  <stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value=""/>
 </launchConfiguration>
